+++ src/app/intro/intro.page.ts
@@ -8,7 +8,7 @@ import { SplashScreen } from '@capacitor/splash-screen';
 import { SettingsService } from '../settings.service';
 import { FavoritesService } from '../favorites.service';
 import { MessageComponent } from '../message/message.component';
-import { addDays, daysBetween, now } from '../utils';
+import { addDays, daysBetween, delay, isWhiteSpace, now } from '../utils';
 import { Dataset } from '../models';
 import { datasetFilename } from '../api';
 import { ApiService } from '../api.service';
@@ -27,6 +27,7 @@ export class IntroPage implements OnInit {
   ready = true;
   showMessage = false;
   downloading = false;
+  yearSelectedAlready = true;
   cards: Dataset[] = [];
   selected: Dataset | undefined;
   message = '';
@@ -36,29 +37,48 @@ export class IntroPage implements OnInit {
     private fav: FavoritesService, private router: Router) { }
 
   async ngOnInit() {
-    this.cards = await this.loadDatasets();
-    this.selected = this.cards[0];
-    this.save(); // Needed in case user has restarted
-    this.ui.setNavigationBar(ThemePrimaryColor);
-    setTimeout(async () => {
-      if (Capacitor.isNativePlatform()) {
-        await StatusBar.setStyle({ style: Style.Dark });
-        await this.ui.setStatusBarBackgroundColor();
-        await SplashScreen.hide();
-        setTimeout(async () => {
-          await this.ui.setStatusBarBackgroundColor();
-        }, 200);
-      }
-    }, 500);
+    // Whether the user has selected a year previously
+    this.yearSelectedAlready = !isWhiteSpace(this.settingsService.settings.dataset);
 
-    setTimeout(async () => {
-      try {
-        this.downloading = true;
-        await this.api.download();
-      } finally {
-        this.downloading = false;
-      }
-    }, 1000);
+    this.cards = await this.loadDatasets();
+
+
+    this.ui.setNavigationBar(ThemePrimaryColor);
+    await delay(500);
+    if (Capacitor.isNativePlatform()) {
+      await StatusBar.setStyle({ style: Style.Dark });
+      await this.ui.setStatusBarBackgroundColor();
+      await SplashScreen.hide();
+      await delay(200);
+      await this.ui.setStatusBarBackgroundColor();
+    }
+
+    // need to load
+    this.load();
+    
+    try {
+      this.downloading = true;
+      await this.api.download();
+    } finally {
+      this.downloading = false;
+    }
+    if (this.yearSelectedAlready) {
+      this.go();
+    }
+  }
+
+  private load() {
+    const idx = this.cards.findIndex((c) => datasetFilename(c) == this.settingsService.settings.dataset);
+
+    if (idx >= 0) {
+      this.selected = this.cards[idx];     
+      console.log('intro.load', this.selected); 
+    } else {
+      // First time in: select this year
+      this.selected = this.cards[0];
+      this.save();
+    }
+    
   }
 
   ionViewWillEnter() {
@@ -73,13 +93,14 @@ export class IntroPage implements OnInit {
     const x = daysBetween(now(), manBurns);
     const until = daysBetween(now(), start);
 
+
+    const hideLocations = (thisYear && until > 1);
+    this.db.setHideLocations(hideLocations);
     //console.log(start, manBurns, x, until);
-    if (thisYear && until > 1) {
+    if (hideLocations && !this.yearSelectedAlready) {
       this.message = `Locations for camps and art will be released in the app on Sunday 27th. There are ${x} days until the man burns.`;
       this.showMessage = true;
-      this.db.setHideLocations(true);
     } else {
-      this.db.setHideLocations(false);
       this.launch();
     }
   }
@@ -96,7 +117,7 @@ export class IntroPage implements OnInit {
       if (Capacitor.isNativePlatform()) {
         setTimeout(async () => {
           this.ui.setNavigationBar();
-          this.ui.hideNavigationBar();
+          //this.ui.hideNavigationBar();
           StatusBar.setStyle({ style: this.ui.darkMode() ? Style.Dark : Style.Light });
           this.ui.setStatusBarBackgroundColor(this.ui.darkMode() ? '#000000' : '#FFFFFF');
         }, 100);
@@ -116,6 +137,7 @@ export class IntroPage implements OnInit {
 
   save() {
     this.settingsService.settings.dataset = datasetFilename(this.selected!);
+    this.settingsService.settings.eventTitle = this.selected!.title;
     this.settingsService.save();
   }