+++ src/app/tabs/tabs.routes.ts
@@ -34,7 +34,11 @@ export const routes: Routes = [
         loadComponent: () => import('../profile/profile.page').then(m => m.ProfilePage)
       },
       {
-        path: 'profile/restrooms',
+        path: 'profile/:mapType',
+        loadComponent: () => import('../pin-map/pin-map.page').then(m => m.PinMapPage)
+      },
+      {
+        path: 'profile/ice',
         loadComponent: () => import('../pin-map/pin-map.page').then(m => m.PinMapPage)
       },
       {