+++ changelog.md
@@ -1,4 +1,14 @@
 ## Changes
+### 1.6
+- Share Art
+- Share Event
+- Share Theme Camp
+- Friends feature
+- Map for Medical
+- Map for Ice
+- Fix padding near year label
+- Fix navigation bar being hidden on Android 12
+- Fix page title animation
 
 ### 1.5
 - Fix for region with non-US date formats